Update Word_Space.md
Changed the way the sentences refer to the 'too tight' and 'too loose' and 'well-balanced' examples. The layout was confusing, and somewhat misleading. I've clarified what is being referred to, by placing the referring line BEFORE each example, and making the position of the ellipsis in each sentence the same.

Line 27: That's an em-dash, not an en-dash. En-dashes are used in number ranges, not as parenthetical dashes.

Line 31: Changed the semi-colon to a comma, as the clause following is a dependent clause. Changed the word 'error' to 'err'.