Update Spacing_Metrics_and_Kerning
Export Peter LeFanu Lumsdaine's typo fix from 257 to other language pages.

Removed dead URL links, first URL is from 2014 and is missing images so it
is not useful (as mentioned by someone in the comment section), while the
second URL has a working javascript, but the other pages are dead, so it's
safer to not suggest to the javascript page as it is unmaintained and is a
possible security risk if replaced in future with malicious code.

Converted menu instructions and keyboard instructions to use same type of
formatting done with earlier fixed pages to highlight instructions and to
maintain consistency/theming.

Imported added improvements from ja-JP into all other language page.

Revert to <ul><li></li></ul> for zh-CN since the bullet points don't seem
to work as well using '-'

Moved French images into fr-FR/images since they are specific to French,
and we don't need to import them into the English pdf (later).
continued... Removed duplicate fa-IR images and use default images.